requirements.txt
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.44.0
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-livereload-server==0.3.2
idna==2.10
jedi==0.17.2
oauthlib==3.1.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
zope.interface==5.2.0
